Add `bytesToZeroOutFromTheBeginning` option
This option, as the name implies, allows the user to declare that a
certain amount of bytes should be zeroed out before starting the
bmap-assisted write process.

The reason for such a task is that in some cases, `bmap-tools` interpret
zero bytes from the head of the image as holes, causing the bmap flash
program to not explicitly write them, and end up confusing some ROM
boot-loaders.

This is really an issue that should be fixed on the bmap generation
phase, but this is a good enough workaround until the issue is dealt
with property.

See: https://github.com/resin-io/etcher/issues/673
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>